+++ packages/blurText/src/index.ts
@@ -1,0 +1,3 @@
+export { default } from '../../../src/components/blurText';
+export { default as BlurText } from '../../../src/components/blurText';
+export type * from '../../../src/components/blurText';