+++ packages/circularText/src/index.ts
@@ -1,0 +1,3 @@
+export { default } from '../../../src/components/circularText';
+export { default as CircularText } from '../../../src/components/circularText';
+export type * from '../../../src/components/circularText';