+++ packages/fuzzyText/src/index.ts
@@ -1,0 +1,3 @@
+export { default } from '../../../src/components/fuzzyText';
+export { default as FuzzyText } from '../../../src/components/fuzzyText';
+export type * from '../../../src/components/fuzzyText';